# Conflicts:
#	android_auth_lib/src/main/java/com/blue_unicorn/android_auth_lib/ctap2/exceptions/StatusCodeException.java

+++ android_auth_lib/src/main/java/com/blue_unicorn/android_auth_lib/ctap2/message_encoding/data/Command.java
@@ -6,7 +6,7 @@ package com.blue_unicorn.android_auth_lib.ctap2.message_encoding.data;
  */
 public interface Command {
 
-    byte getValue();
+    int getValue();
 
     String getParameters();
